# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bessel API",
    "summary": "Bessel HTTP and Webhooks API",
    "description": "Hello! This is the Bessel API.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.getbessel.com",
      "description": "Production environment",
      "x-speakeasy-server-id": "production"
    }
  ],
  "paths": {
...
    "/v1/activity/batch": {
      "post": {
        "tags": [
          "activity"
        ],
        "summary": "Ingest Activity Events",
        "operationId": "ingest_activity_batch_v1_activity_batch_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activity/sources": {
      "get": {
        "tags": [
          "activity"
        ],
        "summary": "List Activity Sources",
        "operationId": "list_activity_sources_v1_activity_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivitySourcesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/activity/summary": {
      "get": {
        "tags": [
          "activity"
        ],
        "summary": "Get Activity Summary",
        "operationId": "get_activity_summary_v1_activity_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Start of time window (Unix epoch seconds, inclusive).",
              "title": "Start Ts"
            },
            "description": "Start of time window (Unix epoch seconds, inclusive)."
          },
          {
            "name": "end_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "End of time window (Unix epoch seconds, exclusive).",
              "title": "End Ts"
            },
            "description": "End of time window (Unix epoch seconds, exclusive)."
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Machine source to query.",
              "title": "Source"
            },
            "description": "Machine source to query."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivitySummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activity/daily": {
      "get": {
        "tags": [
          "activity"
        ],
        "summary": "Get Daily Activity Totals",
        "operationId": "get_daily_activity_v1_activity_daily_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Start of range (Unix epoch seconds, inclusive).",
              "title": "Start Ts"
            },
            "description": "Start of range (Unix epoch seconds, inclusive)."
          },
          {
            "name": "end_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "End of range (Unix epoch seconds, exclusive).",
              "title": "End Ts"
            },
            "description": "End of range (Unix epoch seconds, exclusive)."
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Machine source to query.",
              "title": "Source"
            },
            "description": "Machine source to query."
          },
          {
            "name": "tz_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. 'Europe/Stockholm'). Preferred over tz_offset_mins; handles DST correctly.",
              "title": "Tz Name"
            },
            "description": "IANA timezone name (e.g. 'Europe/Stockholm'). Preferred over tz_offset_mins; handles DST correctly."
          },
          {
            "name": "tz_offset_mins",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Fallback UTC offset in minutes when tz_name is not provided (e.g. 120 for UTC+2). Does not handle DST.",
              "default": 0,
              "title": "Tz Offset Mins"
            },
            "description": "Fallback UTC offset in minutes when tz_name is not provided (e.g. 120 for UTC+2). Does not handle DST."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDailyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activity/intraday": {
      "get": {
        "tags": [
          "activity"
        ],
        "summary": "Get Intraday Activity",
        "operationId": "get_intraday_activity_v1_activity_intraday_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Start of window (Unix epoch seconds, inclusive).",
              "title": "Start Ts"
            },
            "description": "Start of window (Unix epoch seconds, inclusive)."
          },
          {
            "name": "end_ts",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "End of window (Unix epoch seconds, exclusive).",
              "title": "End Ts"
            },
            "description": "End of window (Unix epoch seconds, exclusive)."
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Machine source to query.",
              "title": "Source"
            },
            "description": "Machine source to query."
          },
          {
            "name": "bucket_mins",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "description": "Bucket size in minutes.",
              "default": 15,
              "title": "Bucket Mins"
            },
            "description": "Bucket size in minutes."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityIntradayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ingest-tokens": {
      "get": {
        "tags": [
          "ingest-tokens"
        ],
        "summary": "List Ingest Tokens",
        "operationId": "list_ingest_tokens_v1_ingest_tokens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngestTokenSchema"
                  },
                  "type": "array",
                  "title": "Response List Ingest Tokens V1 Ingest Tokens Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "ingest-tokens"
        ],
        "summary": "Create Ingest Token",
        "operationId": "create_ingest_token_v1_ingest_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestTokenCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/ingest-tokens/{token_id}": {
      "delete": {
        "tags": [
          "ingest-tokens"
        ],
        "summary": "Revoke Ingest Token",
        "operationId": "revoke_ingest_token_v1_ingest_tokens__token_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityAppSummary": {
        "properties": {
          "app_class": {
            "type": "string",
            "title": "App Class",
            "description": "Window class / app identifier."
          },
          "active_secs": {
            "type": "integer",
            "title": "Active Secs",
            "description": "Total active seconds attributed to this app."
          },
          "percentage": {
            "type": "number",
            "title": "Percentage",
            "description": "Share of total active time (0\u2013100)."
          }
        },
        "type": "object",
        "required": [
          "app_class",
          "active_secs",
          "percentage"
        ],
        "title": "ActivityAppSummary"
      },
      "ActivityBatchRequest": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Machine identifier, e.g. hostname."
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActivityEventIn"
            },
            "type": "array",
            "title": "Events",
            "description": "Batch of events to ingest."
          }
        },
        "type": "object",
        "required": [
          "source",
          "events"
        ],
        "title": "ActivityBatchRequest"
      },
      "ActivityBatchResponse": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted",
            "description": "Number of new events inserted."
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "description": "Number of duplicate events skipped."
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "skipped"
        ],
        "title": "ActivityBatchResponse"
      },
      "ActivityDailyEntry": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date",
            "description": "ISO date string (YYYY-MM-DD)."
          },
          "active_secs": {
            "type": "integer",
            "title": "Active Secs",
            "description": "Total active seconds for the day."
          }
        },
        "type": "object",
        "required": [
          "date",
          "active_secs"
        ],
        "title": "ActivityDailyEntry"
      },
      "ActivityDailyResponse": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/ActivityDailyEntry"
            },
            "type": "array",
            "title": "Days",
            "description": "Per-day totals, sorted ascending by date. Days with no activity are omitted."
          }
        },
        "type": "object",
        "required": [
          "days"
        ],
        "title": "ActivityDailyResponse"
      },
      "ActivityEventIn": {
        "properties": {
          "local_id": {
            "type": "integer",
            "title": "Local Id",
            "description": "Original SQLite id from the source machine."
          },
          "ts": {
            "type": "integer",
            "title": "Ts",
            "description": "Unix epoch seconds (UTC) when the event was recorded."
          },
          "state": {
            "type": "string",
            "title": "State",
            "description": "'active' or 'idle'."
          },
          "app_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Class",
            "description": "Window class / app identifier."
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Window title."
          },
          "workspace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace",
            "description": "Workspace name or number."
          }
        },
        "type": "object",
        "required": [
          "local_id",
          "ts",
          "state"
        ],
        "title": "ActivityEventIn"
      },
      "ActivityIntradayBucket": {
        "properties": {
          "bucket": {
            "type": "integer",
            "title": "Bucket",
            "description": "Bucket index from midnight (0-based). Each bucket spans bucket_mins minutes."
          },
          "active_secs": {
            "type": "integer",
            "title": "Active Secs",
            "description": "Total active seconds in this bucket."
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "active_secs"
        ],
        "title": "ActivityIntradayBucket"
      },
      "ActivityIntradayResponse": {
        "properties": {
          "bucket_mins": {
            "type": "integer",
            "title": "Bucket Mins",
            "description": "Duration of each bucket in minutes."
          },
          "total_buckets": {
            "type": "integer",
            "title": "Total Buckets",
            "description": "Total number of buckets covering the requested window."
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/ActivityIntradayBucket"
            },
            "type": "array",
            "title": "Buckets",
            "description": "Buckets with activity, sorted ascending. Buckets with no activity are omitted."
          }
        },
        "type": "object",
        "required": [
          "bucket_mins",
          "total_buckets",
          "buckets"
        ],
        "title": "ActivityIntradayResponse"
      },
      "ActivitySourcesResponse": {
        "properties": {
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "All known machine sources, ordered by most recent activity."
          }
        },
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "ActivitySourcesResponse"
      },
      "ActivitySummaryResponse": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Machine source that was queried."
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "All known machine sources, ordered by most recent activity."
          },
          "total_active_secs": {
            "type": "integer",
            "title": "Total Active Secs",
            "description": "Total active seconds in the requested time window."
          },
          "apps": {
            "items": {
              "$ref": "#/components/schemas/ActivityAppSummary"
            },
            "type": "array",
            "title": "Apps",
            "description": "Per-app breakdown, sorted by active time descending."
          }
        },
        "type": "object",
        "required": [
          "source",
          "sources",
          "total_active_secs",
          "apps"
        ],
        "title": "ActivitySummaryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestTokenCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Identifies the daemon and machine, e.g. 'monitor:laptop'. Creating a token revokes earlier tokens with the same name."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IngestTokenCreate"
      },
      "IngestTokenCreated": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp of the object."
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At",
            "description": "Last modification timestamp of the object."
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "description": "The ID of the object."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "The secret. Returned only once; store it in the daemon's env file."
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "modified_at",
          "id",
          "name",
          "last_used_at",
          "token"
        ],
        "title": "IngestTokenCreated"
      },
      "IngestTokenSchema": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp of the object."
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At",
            "description": "Last modification timestamp of the object."
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "description": "The ID of the object."
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "modified_at",
          "id",
          "name",
          "last_used_at"
        ],
        "title": "IngestTokenSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}